Move checks to pkg
