+++ dynamicbeat/pkg/checks/dns/dns.go
