+++ dynamicbeat/pkg/checks/smb/smb.go
